refactor(/position): 위치정보와 관련된 클래스들을 한 패키지로 모음
