+++ src/main/java/domain/Board.java
@@ -1,5 +1,7 @@
 package domain;
 
+import domain.position.Position;
+
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
